requirements.txt
altair==4.2.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
entrypoints==0.4
geographiclib==2.0
geopy==2.3.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
jsonschema==4.17.3
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.24.3
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==3.20.3
pyarrow==12.0.0
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-docx==0.8.11
pytz==2023.3
requests==2.31.0
rich==13.4.1
semver==3.0.0
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
streamlit==1.12.0
streamlit-aggrid==0.3.4.post3
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
typing-extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.2
validators==0.20.0
watchdog==3.0.0
zipp==3.15.0
